+++ Changelog.md
@@ -1,3 +1,7 @@
+# 0.8.0
+
+* Add parallel workers threads
+
 # 0.7.3
 
 * Add a `--dry-run` option to perform a trial run with no changes made